+++ src/scripts/helpers/memoize-fetch.ts
@@ -1,7 +1,20 @@
-const cachedCalls: Map<string, Promise<Response>> = new Map();
+import { MemoizedFetch, MemoizedFetchOpts } from '../models/memoized-fetch';
+import { fetchHelper } from './fetch';
+import cache from '../../cache'
 
-export default (url: string) : Promise<Response> => {
-  if (cachedCalls.has(url)) return cachedCalls.get(url);
-  cachedCalls.set(url, fetch(url));
-  return cachedCalls.get(url);
+const cachedCalls: Map<string, MemoizedFetch> = cache.memoize;
+
+export default async(url: string, fetchOpts = {}, memoizeOpts: MemoizedFetchOpts = { deleteAfter: { value: 24, period: 'hours' } }): Promise<string> => {
+  const cachedRequest = cachedCalls.get(url);
+
+  if (cachedRequest && (new Date() < new Date(cachedRequest.expiryTime))) {
+    return cachedRequest.response;
+  }
+
+  const response = await fetchHelper(url, fetchOpts);
+  const body = await response.text();
+
+  cachedCalls.set(url, new MemoizedFetch(memoizeOpts).setResponse(body));
+  cache.memoize = cachedCalls;
+  return body;
 };